requirements.txt
absl-py
appdirs
appnope
asn1crypto
astor
attrs
Automat
backcall
bleach
certifi
cffi
constantly
cryptography
decorator
entrypoints
gast
grpcio
html5lib
hyperlink
idna
incremental
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
Markdown
# MarkupSafe==1.0
mistune
nbconvert
nbformat
notebook
numpy
pandocfilters
parso
pexpect
pickleshare
prometheus-client
prompt-toolkit
protobuf
ptyprocess
pyasn1
pyasn1-modules
pycparser
Pygments
pyOpenSSL
python-dateutil
pyzmq
qtconsole
Send2Trash
service-identity
simplegeneric
six
tensorboard
# tensorflow
termcolor
terminado
testpath
tornado
traitlets
Twisted
wcwidth
webencodings
Werkzeug
widgetsnbextension
zope.interface
Django
Pillow
psycopg2-binary